Move mouse behavior to own file
Also enhance mouse behavior. Most important tool for me is the terminal,
so rightclicking the desktop will spawn a new terminal window.
Leftclicking stays as it is, and middle click will open an overview
with all open windows. This way, after middelclicking, one can scroll
through the list and easily get the window in front that one needs.