Add solution to missing function keys
As zsh does not use /etc/inputrc to get the commands sent by the
terminal editor, one would need to find the key codes depending
on system and terminal editor and set them individually per system.
As this is quite cumbersome, with this solution, whenever zsh runs
on a new terminal, it will check if it has a file for it in ~/.zkbd/
and if it doesn't it will query the user for different keys to press
so this can be set up accodingly.

Once set, the keys can be accessed via variables, making it possible
to remove all keycodes in the remaining files and thus making the
files truely portable