Refactor files to use -p switch for which
-p switch for zsh builtin which is similar to the --skip-alias
switch in /usr/bin/which, i.e. if which finds an alias poiting to
any arbitrary binary it is ignored.

Thus which -p less will point to the binary installation of less
(e.g. /usr/bin/less) even though less might be aliased to point to
something else, e.g. alias less=vimpager.

As one does not know if an alias is set anywhere in the config files
but wants to point to the binary, this makes sense for all config files